Add scoring pseudo-code for fast deliberation.
This was entertaining, but it's perhaps wrong,
although we can't find any reason why so far.
We contacted the world's top specialist on the matter.
…

Three gods, A, B, and C are called,
in no particular order, True, False, and Random.

True always speaks truly, False always speaks falsely,
but whether Random speaks truly or falsely is a
completely random matter.

Your task is to determine the identities of A, B, and C
by asking three yes-no questions; each question
must be put to exactly one god. The gods understand
English, but will answer all questions in
their own language, in which the words for yes
and no are da and ja, in some order.

You do not know which word means which.